// src/infrastructure/external-services/ai-generate-image/comfyui/control-net/types/ControlnetNameEnum.ts
export enum ControlnetNameEnum {
  T2IADAPTER_DEPTH = 't2iadapter_depth_sd14v1.pth',
  OPEN_POSE = 'control_v11p_sd15_openpose_fp16.safetensors',
  SKETCH = '',
  COLOR = '',
  T2IADAPTER_LINEAR_SDXL = 't2i-adapter-lineart-sdxl-1.0.safetensors',
  T2IADAPTER_CANNY_SDXL = 't2i-adapter-canny-sdxl-1.0.safetensors',
  T2IADAPTER_DEPTH_SDXL = 't2i-adapter-depth-midas-sdxl-1.0.safetensors',
  T2IADAPTER_SKETCH_SDXL = 't2i-adapter-sketch-sdxl-1.0.safetensors',
  T2IADAPTER_OPENPOSE_SDXL = 't2i-adapter-openpose-sdxl-1.0.safetensors',
}